chore: use ProfileBox to calculate connected state
