+++ src/extension/options-page/DashboardComponents/ProfileBox.tsx
@@ -1,0 +1,81 @@
+import React from 'react'
+import { Persona } from '../../../database'
+import { definedSocialNetworkWorkers } from '../../../social-network/worker'
+
+import classNames from 'classnames'
+import ProviderLine from './ProviderLine'
+import { makeStyles } from '@material-ui/styles'
+import { createStyles, Typography, Theme } from '@material-ui/core'
+import { geti18nString } from '../../../utils/i18n'
+import { useColorProvider } from '../../../utils/theme'
+
+const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        line: {
+            display: 'flex',
+            alignItems: 'center',
+            '&:not(:first-child)': {
+                paddingTop: theme.spacing(1),
+            },
+            '& > div': {
+                overflow: 'hidden',
+                textOverflow: 'ellipsis',
+                flexShrink: 1,
+                whiteSpace: 'nowrap',
+            },
+            '& > .content': {
+                margin: '0 1em',
+            },
+            '& > .title': {
+                flexShrink: 0,
+                width: '5rem',
+            },
+            '& > .extra-item': {
+                visibility: 'hidden',
+                marginRight: '0',
+                flexShrink: 0,
+                marginLeft: 'auto',
+                cursor: 'pointer',
+                fontSize: '0.8rem',
+            },
+            '&:hover': {
+                '& > .extra-item': {
+                    visibility: 'visible',
+                },
+            },
+        },
+    }),
+)
+
+interface Props {
+    persona: Persona | null
+    border?: true
+}
+
+export default function ProfileBox({ persona, border }: Props) {
+    const classes = useStyles()
+    const color = useColorProvider()
+    const profiles = persona ? [...persona.linkedProfiles] : []
+
+    const providers = [...definedSocialNetworkWorkers].map(i => {
+        const profile = profiles.find(([key, value]) => key.network === i.networkIdentifier)
+        return {
+            network: i.networkIdentifier,
+            connected: !!profile,
+            id: profile?.[0].userId,
+        }
+    })
+
+    return (
+        <>
+            {providers.map(provider => (
+                <Typography className={classes.line} component="div">
+                    <ProviderLine {...provider} border={border ?? false}></ProviderLine>
+                    {provider.connected && (
+                        <div className={classNames('extra-item', color.error)}>{geti18nString('disconnect')}</div>
+                    )}
+                </Typography>
+            ))}
+        </>
+    )
+}
